activeRunning
added check if cmd is active, if not than it won't be performed. Added also a checkbox that is disabled but updated programmatically to keep track if there is a thread still running that command. In case button is pressed than all command will run except those not active or those that there is a thread still running that command